Also split the GL line stuff
